PeopleDB/src/test/java/yp/peopledb/repository/PeopleRepositoryTest.java
package yp.peopledb.repository;

import org.junit.jupiter.api.AfterEach;
import org.junit.jupiter.api.BeforeEach;
import org.junit.jupiter.api.Disabled;
import org.junit.jupiter.api.Test;
import yp.peopledb.model.Person;

import java.sql.Connection;
import java.sql.DriverManager;
import java.sql.SQLException;
import java.time.ZoneId;
import java.time.ZonedDateTime;
import java.util.Arrays;
import java.util.Optional;

import static java.util.stream.Collectors.joining;
import static org.assertj.core.api.Assertions.assertThat;

public class PeopleRepositoryTest {

    private Connection connection;
    private PeopleRepository repo;

    @BeforeEach
    void setUp() throws SQLException {
        connection = DriverManager.getConnection(
                "jdbc:h2:~/peopletest".replace("~", System.getProperty("user.home"))
        );
        connection.setAutoCommit(false);
        repo = new PeopleRepository(connection);
    }

    @AfterEach
    void tearDown() throws SQLException {
        if (connection != null) {
            connection.close();
        }
    }

    @Test
    @Disabled("This test is failing on GitHub")
    public void canSaveOnePerson() {
        Person john = new Person("John", "Smith",
                ZonedDateTime.of(1980, 11, 15 ,
                        15, 15, 0, 0,
                        ZoneId.of("-6")));
        Person savedPerson = repo.save(john);
        assertThat(savedPerson.getId()).isGreaterThan(0);
    }
    @Test
    @Disabled("This test is failing on GitHub")
    public void canSaveTwoPersons(){
        Person john = new Person("John", "Smith",
                ZonedDateTime.of(1980, 11, 15 ,
                        15, 15, 0, 0,
                        ZoneId.of("-6")));
        Person bobby = new Person("Bobby", "Smith",
                ZonedDateTime.of(1982, 9, 13 ,
                        15, 15, 0, 0,
                        ZoneId.of("-6")));
        Person savedPerson1 = repo.save(john);
        Person savedPerson2 = repo.save(bobby);
        assertThat(savedPerson1.getId()).isNotEqualTo(savedPerson2.getId());
    }

    @Test
    @Disabled("This test is failing on GitHub")
    public void canFindPersonById (){
        Person savedPerson = repo.save(new Person("test", "jackson", ZonedDateTime.now()));
        Person foundPerson = repo.findById(savedPerson.getId()).get();
        assertThat(foundPerson).isEqualTo(savedPerson);
    }

    @Test
    @Disabled("This test is failing on GitHub")
    public void testPersonIdNotFound (){
        Optional<Person> foundPerson = repo.findById(-1L);
        assertThat(foundPerson).isEmpty();
    }

    @Test
    @Disabled("This test is failing on GitHub")
    public void canGetCount (){
        long startCount = repo.count();
        repo.save(new Person("John1", "Smith",
                ZonedDateTime.of(1980, 11, 15 ,
                        15, 15, 0, 0,
                        ZoneId.of("-6"))));
        long endCount = repo.count();
        assertThat(endCount).isEqualTo(startCount + 1);
    }

    @Test
    @Disabled("This test is failing on GitHub")
    public void canDelete (){
        Person savedPerson = new Person("SavedPerson", "Smith",
                ZonedDateTime.of(1980, 11, 15 ,
                        15, 15, 0, 0,
                        ZoneId.of("-6")));
        repo.save(savedPerson);
        long startCount = repo.count();
        repo.delete(savedPerson);
        long endCount = repo.count();
        assertThat(endCount).isEqualTo(startCount - 1);
    }

    @Test
    @Disabled("This test is failing on GitHub")
    public void canDeleteMultiplePeople (){
        Person p1 = new Person("P1", "Smith",
                ZonedDateTime.of(1980, 11, 15 ,
                        15, 15, 0, 0,
                        ZoneId.of("-6")));
        Person p2 = new Person("P2", "Smith",
                ZonedDateTime.of(1980, 11, 15 ,
                        15, 15, 0, 0,
                        ZoneId.of("-6")));
        repo.save(p1);
        repo.save(p2);
        long startCount = repo.count();
        repo.delete(p1, p2);
        long endCount = repo.count();
        assertThat(endCount).isEqualTo(startCount - 2);
    }

    @Test
    public void experiment(){
        Person p1 = new Person( 10L, null, null, null);
        Person p2 = new Person( 20L, null, null, null);
        Person p3 = new Person( 30L, null, null, null);
        Person p4 = new Person( 40L, null, null, null);

        Person[] people = Arrays.asList(p1, p2, p3, p4).toArray(new Person[]{});
        String ids = Arrays.stream(people).
                map(Person::getId).
                map(String::valueOf).
                collect(joining(","));
        System.out.println(ids);

    }
}